requirements.txt
# Faster Whisper para transcripción de audio
faster-whisper==1.0.3

# Servidor web Flask para API
Flask==3.0.0
Flask-CORS==4.0.0

# Procesamiento de audio
librosa==0.10.1
soundfile==0.12.1

# Utilidades
requests==2.31.0
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0

# Logging y configuración
python-dotenv==1.0.0
loguru==0.7.2

# Validación de archivos
filetype==1.2.0